requirements.txt
torch>=1.0
torchvision
pytorch-ignite
visdom
pytest
tqdm
numpy
scikit-learn
pyyaml
opencv-python
gulpio